ml/requirement.txt
flask
ultralytics
opencv-python
numpy
pillow
requests
pandas
scikit-learn
matplotlib
torch
torchvision
